adjusting tests and moving duplicate to penalty response
